refactor(game): refactor used types into files
Refactor vector operations and declarations into separate files
